third_party/WebKit/LayoutTests/fast/html/imports/import-opening-inspector-expected.txt

PASS window.greeting is 'Hello, World!'
PASS successfullyParsed is true

TEST COMPLETE
